add fosjsrouting, translate words to lithuanian language

+++ src/DTR/DTRBundle/Controller/CartController.php
@@ -15,6 +15,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Component\HttpFoundation\Response;
 use Sensio\Bundle\FrameworkExtraBundle\Request\ParamConverter;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 
 class CartController extends Controller
 {
@@ -51,45 +52,76 @@ class CartController extends Controller
     }
 
     /**
-     * @Route("/event/{hash}/shops/{shopName}/remove-from-cart/{itemId}", name="_remove_from_cart")
+     * @Route("/event/{hash}/remove-from-cart/{itemId}", name="_remove_from_cart")
      * @Template()
      */
-    public function removeFromCartAction($itemId, $hash, $shopName)
+    public function removeFromCartAction(Request $request, $itemId, $hash)
     {
-        $em = $this->getDoctrine()->getManager();
-        $item = $em->getRepository('DTRBundle:Item')->find($itemId);
-        $member = $item->getMember();
-        $member->removeItem($item);
-        $em->remove($item);
+        if($request->isXmlHttpRequest()) {
+            return "hello";
+        } else {
+            $em = $this->getDoctrine()->getManager();
+            $event = $em->getRepository('DTRBundle:Event')->findByHash($hash);
+            $item = $em->getRepository('DTRBundle:Item')->find($itemId);
+            $shopName = $item->getProduct()->getShop()->getSlug();
+            $member = $item->getMember();
+            $member->removeItem($item);
+            $em->remove($item);
 
-        $em->flush();
+            $em->flush();
 
-        return $this->forward('DTRBundle:Default:shop', array(
-            'hash' => $hash,
-            'shopName' => $shopName
-        ));
+            return $this->redirectToRoute('_shop', ['hash' => $event[0]->getHash(), 'shopName' => $shopName]);
+        }
     }
 
     /**
-     * @Route("/event/{hash}/shops/{shopName}/delete-cart", name="_delete_cart")
+     * @Route("/event/{hash}/remove-from-cart/{itemId}", name="_remove_from_cart_ajax")
+     * @Template()
+     * @Method({"POST"})
+     */
+    public function removeFromCartAjaxAction(Request $request, $itemId, $hash)
+    {
+        if($request->isXmlHttpRequest()) {
+            return new Response("HELLO!");
+        } else {
+            $em = $this->getDoctrine()->getManager();
+            $event = $em->getRepository('DTRBundle:Event')->findByHash($hash);
+            $item = $em->getRepository('DTRBundle:Item')->find($itemId);
+            $shopName = $item->getProduct()->getShop()->getSlug();
+            $member = $item->getMember();
+            $member->removeItem($item);
+            $em->remove($item);
+
+            $em->flush();
+
+            return $this->redirectToRoute('_shop', ['hash' => $event[0]->getHash(), 'shopName' => $shopName]);
+        }
+    }
+
+    /**
+     * @Route("/event/{hash}/delete-cart", name="_delete_cart")
      * @Template()
      */
-    public function deleteCartAction($hash, $shopName)
+    public function deleteCartAction(Request $request, $hash)
     {
-        $em = $this->getDoctrine()->getManager();
-        $user = $this->getUser();
-        $event = $em->getRepository('DTRBundle:Event')->findByHash($hash);
-        $member = $em->getRepository('DTRBundle:Member')->findByEventUser($event[0], $user);
-        $items = $em->getRepository('DTRBundle:Item')->findByMember($member);
-        foreach ($items as $item) {
-            $em->remove($item);
-        }
-        $em->flush();
+        if($request->isXmlHttpRequest()) {
+            return new Response("HELLO!");
+        } else {
+            $em = $this->getDoctrine()->getManager();
+            $user = $this->getUser();
+            $event = $em->getRepository('DTRBundle:Event')->findByHash($hash);
+            $member = $em->getRepository('DTRBundle:Member')->findByEventUser($event[0], $user);
+            $items = $em->getRepository('DTRBundle:Item')->findByMember($member);
+            $shopName = $items[0]->getProduct()->getShop()->getSlug();
+            foreach ($items as $item) {
+                $em->remove($item);
+            }
+            $em->flush();
 
-        return $this->forward('DTRBundle:Default:shop', array(
-            'hash' => $hash,
-            'shopName' => $shopName
-        ));
+
+            return $this->redirectToRoute('_shop', ['hash' => $event[0]->getHash(), 'shopName' => $shopName]);
+        }
+
 
     }
